:recycle: Refactor Home module
- Refactor into Album and Hero Components
- Add version to Top Menu
