+++ src/App.tsx
@@ -2,19 +2,22 @@ import React from 'react';
 
 import { Provider } from 'react-redux'
 import { createStore, applyMiddleware, Store } from 'redux'
-import {rootReducer, RootState} from './store/reducers'
+import { rootReducer, RootState } from './store/reducers'
 import thunk from "redux-thunk";
 
 import Routes from './routes/routes';
 
-const store : Store <RootState> & {
+import CssBaseline from '@material-ui/core/CssBaseline';
+
+const store: Store<RootState> & {
   dispatch: DispatchType;
-} = createStore(rootReducer , applyMiddleware(thunk))
+} = createStore(rootReducer, applyMiddleware(thunk))
 
 function App() {
 
   return (
     <Provider store={store}>
+      <CssBaseline />
       <Routes />
     </Provider>
   );
