+++ src/components/elements/Copyright/Copyright.tsx
@@ -5,11 +5,11 @@ import React, { ReactElement } from 'react'
 export default function Copyright(): ReactElement {
     return (
         <Typography variant="body2" color="textSecondary" align="center">
-            {'No Copyright ©'}  {new Date().getFullYear()}
+            {'No Copyright © '} 
             <Link color="inherit" href="https://martinreboredo.com/">
                 Martin Reboredo
                 </Link>
-            <Box component="span" display="inline" data-testid="date">
+            <Box component="span" display="inline" data-testid="date" ml={1}>
                 {new Date().getFullYear()}
             </Box>
         </Typography>
