+++ src/components/elements/Hero/Hero.tsx
@@ -1,0 +1,52 @@
+import React from 'react';
+import Button from '@material-ui/core/Button';
+import Grid from '@material-ui/core/Grid';
+import Typography from '@material-ui/core/Typography';
+import { makeStyles } from '@material-ui/core/styles';
+import Container from '@material-ui/core/Container';
+
+
+const useStyles = makeStyles((theme) => ({
+
+  heroContent: {
+    backgroundColor: theme.palette.background.paper,
+    padding: theme.spacing(8, 0, 6),
+  },
+  heroButtons: {
+    marginTop: theme.spacing(4),
+  },
+}));
+
+
+export default function Album() {
+  const classes = useStyles();
+
+  return (
+    <>
+        <div className={classes.heroContent}>
+          <Container maxWidth="sm">
+            <Typography component="h1" variant="h2" align="center" color="textPrimary" gutterBottom>
+              MERN
+            </Typography>
+            <Typography variant="h5" align="center" color="textSecondary" paragraph>
+              Boilerplate Code for a MERN. Using Docker, Redux, Google Cloud, Material-UI.
+            </Typography>
+            <div className={classes.heroButtons}>
+              <Grid container spacing={2} justify="center">
+                <Grid item>
+                  <Button href="https://github.com/martinrebo/boilerplate" target="_blank" variant="contained" color="primary">
+                    See Source Code
+                  </Button>
+                </Grid>
+                <Grid item>
+                  <Button href="https://www.martinreboredo.com/" target="_blank" variant="outlined" color="primary">
+                    Contact Me
+                  </Button>
+                </Grid>
+              </Grid>
+            </div>
+          </Container>
+        </div>
+    </>
+  );
+}
